respaldo de seguridad ms-notificaciones

+++ ms-auditoria/src/main/java/com/fixnow/msauditoria/Repository/AuditoriaRepository.java
@@ -9,12 +9,9 @@ import java.util.List;
 @Repository
 public interface AuditoriaRepository extends JpaRepository<Auditoria, Integer> {
 
-    // Recupera la trazabilidad de las acciones en el contexto de un ticket específico
-    List<Auditoria> findByIdTicket(Integer idTicket);
+    List<Auditoria> findByIdTicket(long idTicket);
 
-    // Recupera la trazabilidad de las acciones asociadas a un pago específico
     List<Auditoria> findByIdPago(Integer idPago);
 
-    // Recupera la trazabilidad de las acciones asociadas a una persona específica
-    List<Auditoria> findByIdPersona(Integer idPersona);
+    List<Auditoria> findByIdPersona(long idPersona);
 }
